Cleanup and rename community to members component
