+++ src/modules/dashboard/components/ColonyMembers/ColonyMembers.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 
-import Community from '~dashboard/Community';
+import Members from '~dashboard/Members';
 import SubscribedColoniesList from '~dashboard/SubscribedColoniesList/SubscribedColoniesList';
 import { useColonyFromNameQuery } from '~data/index';
 import styles from './ColonyMembers.css';
@@ -25,7 +25,7 @@ const ColonyMembers = () => {
       <div className={styles.mainContentGrid}>
         <div className={styles.mainContent}>
           {colonyData && colonyData.colony && (
-            <Community colony={colonyData.colony} />
+            <Members colony={colonyData.colony} />
           )}
         </div>
         <aside className={styles.rightAside}>Edit permissions</aside>
