--- src/modules/dashboard/components/Community/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Community';
