+++ src/modules/dashboard/components/Members/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Members';
